requirements.txt
distro-info
influxdb
launchpadlib
PyYAML
